+++ backend/src/modules/album/album.service.ts
@@ -1,49 +1,77 @@
+import { NotFoundError } from "@/errors/ApiError.js";
 import type { AlbumRepository } from "@/repositories/prisma/album.repository.js";
 import { FileService, FileType } from "@/services/fileService.js";
 import type {
-  addToCollectionType,
-  AlbumRepositoryType,
-  createAlbumType,
-  updateAlbumType,
-} from "@/types/album/album.js";
-import type { AlbumServiceType } from "@/types/album/album.js";
+  AddToCollectionDTO,
+  AlbumsResponse,
+  CreateAlbumDTO,
+  IAlbumRepository,
+  IAlbumService,
+  UpdateAlbumDTO,
+  FormattedAlbum,
+  FormattedDetailedAlbum,
+} from "@/types/album/index.js";
 import type { queryType } from "@/types/common/query.js";
+import {
+  formatAlbum,
+  formatDetailedAlbum,
+} from "@/utils/formatters/album.formatter.js";
 import type { MultipartFile } from "@fastify/multipart";
 
-export class AlbumService implements AlbumServiceType {
+export class AlbumService implements IAlbumService {
   constructor(
-    private albumRepository: AlbumRepositoryType,
+    private albumRepository: IAlbumRepository,
     private fileService: FileService,
   ) {}
 
-  async getAlbum(id: string) {
-    return await this.albumRepository.findById(id);
+  async getAlbum(id: string): Promise<FormattedDetailedAlbum | null> {
+    const album = await this.albumRepository.findById(id);
+    return formatDetailedAlbum(album);
   }
 
-  async getAlbums(data: queryType) {
-    return await this.albumRepository.findAll(data);
+  async getAlbums(options: queryType): Promise<AlbumsResponse> {
+    const { albums, total } = await this.albumRepository.findAll(options);
+    const { count, offset } = options;
+    return {
+      data: albums.map((album) => formatAlbum(album)),
+      meta: {
+        total,
+        count,
+        offset,
+      },
+    };
   }
 
-  async createAlbum(data: createAlbumType, pic: MultipartFile) {
-    let picPath; // если удалить придется
+  async createAlbum(
+    data: CreateAlbumDTO,
+    pic: MultipartFile,
+  ): Promise<FormattedAlbum> {
+    let picPath: string | undefined; // если потребуется откат
     try {
       const picturePath = await this.fileService.createFile(
         FileType.IMAGE,
         pic,
       );
       picPath = picturePath;
-      return await this.albumRepository.create(data, picturePath);
+      const album = await this.albumRepository.create(data, picturePath);
+      return formatAlbum(album);
     } catch (err) {
       if (picPath) await this.fileService.removeFile(picPath);
       throw err;
     }
   }
 
-  async updateAlbum(id: string, data: updateAlbumType, pic?: MultipartFile) {
-    let picPath; // если удалить придется
+  async updateAlbum(
+    id: string,
+    data: UpdateAlbumDTO,
+    pic?: MultipartFile,
+  ): Promise<FormattedAlbum> {
+    let picPath: string | undefined; // если потребуется откат
     try {
       if (pic) {
         const album = await this.albumRepository.findById(id);
+        if (!album)
+          throw new NotFoundError(`Album with id ${id} was not found!`);
         await this.fileService.removeFile(album?.picture);
         const picturePath = await this.fileService.createFile(
           FileType.IMAGE,
@@ -51,18 +79,19 @@ export class AlbumService implements AlbumServiceType {
         );
         picPath = picturePath;
       }
-      return await this.albumRepository.update(id, data, picPath);
+      const album = await this.albumRepository.update(id, data, picPath);
+      return formatAlbum(album);
     } catch (err) {
       if (picPath) await this.fileService.removeFile(picPath);
       throw err;
     }
   }
 
-  async deleteAlbum(id: string) {
-    return await this.albumRepository.remove(id);
+  async deleteAlbum(id: string): Promise<void> {
+    return this.albumRepository.remove(id);
   }
 
-  async addToCollection(data: addToCollectionType) {
-    return await this.albumRepository.addToCollection(data);
+  async addToCollection(data: AddToCollectionDTO): Promise<void> {
+    return this.albumRepository.addToCollection(data);
   }
 }
