+++ backend/src/modules/track/track.service.ts
@@ -1,51 +1,78 @@
+import { NotFoundError } from "@/errors/ApiError.js";
 import { TrackRepository } from "@/repositories/prisma/track.repository.js";
 import { FileType, type FileService } from "@/services/fileService.js";
 import type { queryType } from "@/types/common/query.js";
 import type {
-  addToAlbumType,
-  addToPlaylistType,
-  createTrackType,
-  TrackRepositoryType,
-  TrackServiceType,
-  updateTrackType,
-} from "@/types/track/track.js";
+  ITrackRepository,
+  ITrackService,
+  CreateTrackDTO,
+  UpdateTrackDTO,
+  AddToAlbumDTO,
+  AddToPlaylistDTO,
+  FormattedTrack,
+  FormattedDetailedTrack,
+  TracksResponse,
+} from "@/types/track/index.js";
+import {
+  formatDetailedTrack,
+  formatTrack,
+} from "@/utils/formatters/track.formatter.js";
 import type { MultipartFile } from "@fastify/multipart";
 
-export class TrackService implements TrackServiceType {
+export class TrackService implements ITrackService {
   constructor(
-    private trackRepository: TrackRepositoryType,
+    private trackRepository: ITrackRepository,
     private fileService: FileService,
   ) {}
 
-  async getTrack(id: string) {
-    return this.trackRepository.findById(id);
+  async getTrack(id: string): Promise<FormattedDetailedTrack | null> {
+    const track = await this.trackRepository.findById(id);
+    return formatDetailedTrack(track);
   }
 
-  async getTracks(opts: queryType) {
-    return this.trackRepository.findAll(opts);
+  async getTracks(opts: queryType): Promise<TracksResponse> {
+    const { count, offset } = opts;
+    const { total, tracks } = await this.trackRepository.findAll(opts);
+    return {
+      data: tracks.map((track) => formatTrack(track)),
+      meta: {
+        total,
+        count,
+        offset,
+      },
+    };
   }
 
-  async createTrack(data: createTrackType, audio: MultipartFile) {
-    let audPath; // если удалить придется
+  async createTrack(
+    data: CreateTrackDTO,
+    audio: MultipartFile,
+  ): Promise<FormattedTrack> {
+    let audPath: string | undefined; // если понадобится откатить файл
     try {
       const audioPath = await this.fileService.createFile(
         FileType.AUDIO,
         audio,
       );
       audPath = audioPath;
-      console.log(audPath);
-      return await this.trackRepository.create(data, audioPath);
+      const track = await this.trackRepository.create(data, audioPath);
+      return formatTrack(track);
     } catch (err) {
       if (audPath) await this.fileService.removeFile(audPath);
       throw err;
     }
   }
 
-  async updateTrack(id: string, data: updateTrackType, audio?: MultipartFile) {
-    let audPath; // если удалить придется
+  async updateTrack(
+    id: string,
+    data: UpdateTrackDTO,
+    audio?: MultipartFile,
+  ): Promise<FormattedTrack> {
+    let audPath: string | undefined; // если понадобится откатить файл
     try {
       if (audio) {
         const track = await this.trackRepository.findById(id);
+        if (!track)
+          throw new NotFoundError(`Album with id ${id} was not found!`);
         await this.fileService.removeFile(track?.audio);
         const audioPath = await this.fileService.createFile(
           FileType.AUDIO,
@@ -53,39 +80,23 @@ export class TrackService implements TrackServiceType {
         );
         audPath = audioPath;
       }
-      return await this.trackRepository.update(id, data, audPath);
+      const track = await this.trackRepository.update(id, data, audPath);
+      return formatTrack(track);
     } catch (err) {
       if (audPath) await this.fileService.removeFile(audPath);
       throw err;
     }
   }
 
-  async listenIncrement(id: string) {
+  async listenIncrement(id: string): Promise<void> {
     return this.trackRepository.listenIncrement(id);
   }
 
-  async deleteTrack(id: string) {
+  async deleteTrack(id: string): Promise<void> {
     return this.trackRepository.remove(id);
   }
 
-  async addToAlbum(data: addToAlbumType) {
-    return this.trackRepository.addToAlbum(data);
-  }
-
-  async addToFavorite(trackID: string) {
-    //   return this.trackRepository.addToFavourite(trackID);
-  }
-
-  async addToPlaylist(data: addToPlaylistType) {
+  async addToPlaylist(data: AddToPlaylistDTO): Promise<void> {
     return this.trackRepository.addToPlaylist(data);
-  }
-
-  async getRecommendations(
-    searchQuery = "",
-    count: number = 10,
-    offset: number = 0,
-  ) {
-    console.log(`Getting recommendations: count=${count}, offset=${offset}`);
-    return { recommendations: [], count, offset };
   }
 }
